refactor(entity): split navigation/abilities/movement_validation along reader/manager/wire/bounds seams
Pure code movement to bring three over-cap files under the 700-line hard
cap (CLAUDE.md file-organization). Each X.rs becomes X/mod.rs + submodules;
mod.rs re-exports all public items so every external `use crate::...` path
is unchanged.

navigation.rs (1302) -> navigation/
  - xrc.rs: XRC binary reader helpers + header sanity caps (read_*,
    MAX_*, check_count, checked_alloc_size) + their pure widening/log
    guards.
  - mod.rs: NavMesh struct, query API (load/find_path/raycast/...),
    path-size + search-extent constants, Debug impl.
  - tests.rs: NavMesh load/query integration tests + hostile-`.nav`
    header regression guards (exercise NavMesh::load).

abilities.rs (1236) -> abilities/
  - defs.rs: flag/target/timer/result/damage/stat constants, AbilityDef,
    EffectDef, AbilityTreeData.
  - manager.rs: CooldownEntry, AbilityManager (cooldowns, auto-cycle,
    weapon-grant accounting).
  - wire.rs: ClientEffectResult, serialize_timer_update,
    serialize_effect_results.
  - mod.rs: re-exports.

movement_validation.rs (754) -> movement_validation/
  - bounds.rs: SpaceBounds + position_within_bounds (finite/AABB layer).
  - mod.rs: MovementReject, SpeedSample, KinematicsOutcome,
    MovementValidator (speed/teleport layer), TELEPORT_GRACE.

Added net-new pure-unit coverage for EffectDef::param_i32/param_f32
(present / missing / unparseable). Fixture-gated NavMesh coverage gaps
logged to #531.

Behavior-preserving · re-export discipline · zero public-surface change

Part of #529
