--- crates/entity/src/abilities.rs
@@ -1,1236 +1,0 @@
-//! Ability system for entity combat.
-//!
-//! Each being has an `AbilityManager` that tracks known abilities, cooldowns,
-//! and the currently active ability. Abilities are defined in the database
-//! (`resources.abilities`) and organized into archetype-specific ability trees.
-//!
-//! Reference: `python/cell/AbilityManager.py`, `python/common/defs/Ability.py`
-
-use std::collections::{HashMap, HashSet};
-use std::time::{Duration, Instant};
-
-// ── Ability flags (from python/Atrea/enums.py) ────────────────────────────
-
-pub const AF_USE_WEAPON_RANGE: u32 = 4;
-pub const AF_RESPONSE: u32 = 16;
-pub const AF_DEACTIVATE_AUTO_CYCLE: u32 = 1024;
-pub const AF_SPEED_GRENADE: u32 = 2048;
-pub const AF_SPEED_DEPLOY: u32 = 4096;
-pub const AF_SPEED_ATTACK: u32 = 8192;
-/// Channelled abilities normally cancel when the channeller moves more
-/// than `CHANNEL_INTERRUPT_DISTANCE` from their channel-start position.
-/// Setting this flag exempts the ability — useful for "channels while
-/// running" content (sustained beam abilities the player can walk with).
-///
-/// Default for every authored ability today is 0 (cancel-on-move). Flip
-/// the bit per-ability as content lands that should be movement-tolerant.
-/// Reserved bit 14 — not in any python reference; original game's
-/// canonical name unknown, this is the Cimmeria-side name.
-pub const AF_CHANNEL_ALLOWS_MOVEMENT: u32 = 16384;
-
-// ── Target types ──────────────────────────────────────────────────────────
-
-pub const TARGET_NONE: i32 = 0;
-pub const TARGET_SELF: i32 = 1;
-pub const TARGET_TARGET: i32 = 2;
-pub const TARGET_GROUND: i32 = 3;
-
-// ── Target collection methods (TCM) — per-effect dispatch shape ──────────
-//
-// Authored in the original game's data as string literals. We keep them as
-// strings on `EffectDef` (rather than mapping to an enum) so unknown
-// values from new content surface as a `tracing::warn!` instead of
-// silently coercing to a default. The three values that occur in seed:
-//   - TCM_Single    (2795 rows) — primary target only
-//   - TCM_AERadius  (300 rows)  — every entity within `Radius` of the
-//                                 source/ground point
-//   - TCM_AECone    (99 rows)   — every entity inside a cone anchored at
-//                                 the source, oriented toward the primary
-//                                 target, with length+width tiers from
-//                                 `tcm_param1` and `tcm_param2`
-pub const TCM_SINGLE: &str = "TCM_Single";
-pub const TCM_AE_RADIUS: &str = "TCM_AERadius";
-pub const TCM_AE_CONE: &str = "TCM_AECone";
-
-// ── Effect flags (subset of `resources.effects.flags` bitmask) ───────────
-//
-// Authored as a packed integer per-effect. The bits we honor today:
-pub const EF_INTERRUPT_CHANCE: u32 = 16; // category: interrupt roll
-pub const EF_STUN: u32 = 12; // category: stun on apply (target loses controls)
-pub const EF_DONT_USE_QR: u32 = 32; // category: bypass QR (always-hit)
-pub const EF_MENTAL_RESIST_ROLL: u32 = 64; // category: target rolls resist
-pub const EF_SUPPRESSION: u32 = 76; // category: movement slow + accuracy debuff
-pub const EF_EXTRA_DAMAGE: u32 = 512; // category: bonus damage on second pulse
-pub const EF_DOT: u32 = 516; // category: damage-over-time (pulses)
-
-// ── Timer types (sent via onTimerUpdate) ──────────────────────────────────
-
-pub const TIMER_ABILITY_WARMUP: i8 = 1;
-pub const TIMER_ABILITY_COOLDOWN: i8 = 2;
-pub const TIMER_DURATION_EFFECT: i8 = 5;
-pub const TIMER_CATEGORY_COOLDOWN: i8 = 8;
-
-// ── Result codes ──────────────────────────────────────────────────────────
-
-pub const RC_NONE: u8 = 0;
-pub const RC_HIT: u8 = 1;
-pub const RC_MISS: u8 = 2;
-pub const RC_CRITICAL: u8 = 3;
-pub const RC_DOUBLE_CRITICAL: u8 = 4;
-pub const RC_GLANCING: u8 = 5;
-
-// ── Damage types ──────────────────────────────────────────────────────────
-
-pub const DT_UNTYPED: i8 = 0;
-pub const DT_ENERGY: i8 = 1;
-pub const DT_HAZMAT: i8 = 2;
-pub const DT_PHYSICAL: i8 = 3;
-pub const DT_PSIONIC: i8 = 4;
-
-// ── Stat result codes ─────────────────────────────────────────────────────
-
-pub const SRC_NONE: i8 = 0;
-pub const SRC_ABSORB: i8 = 1;
-pub const SRC_IMMUNE: i8 = 2;
-pub const SRC_MORTAL: i8 = 3;
-
-// ── AbilityDef ────────────────────────────────────────────────────────────
-
-/// An ability definition loaded from the database.
-///
-/// Mirrors `python/common/defs/Ability.py` fields needed for validation and
-/// invocation. Full damage/effect resolution will use the effect system.
-#[derive(Debug, Clone)]
-pub struct AbilityDef {
-    pub ability_id: i32,
-    pub name: String,
-    pub cooldown: f32,
-    pub warmup: f32,
-    pub flags: u32,
-    pub is_ranged: bool,
-    pub min_range: i32,
-    pub max_range: i32,
-    pub target_type_id: i32,
-    pub effect_ids: Vec<i32>,
-    pub moniker_ids: Vec<i64>,
-    pub required_ammo: i32,
-    pub event_set_id: Option<i32>,
-    pub velocity: f32,
-}
-
-/// A single effect within an ability (damage, heal, buff, etc).
-///
-/// Mirrors the columns on `resources.effects` plus the NVP bag from
-/// `resources.effect_nvps`. The dispatch-relevant fields (`pulse_count`,
-/// `pulse_duration`, `is_channeled`, `target_collection_method`,
-/// `tcm_param1/2`, `flags`) were added to support pulsing, cone AoE, and
-/// flag-driven script categories.
-#[derive(Debug, Clone)]
-pub struct EffectDef {
-    pub effect_id: i32,
-    pub ability_id: i32,
-    pub delay: i32,
-    pub effect_sequence: i32,
-    pub event_set_id: Option<i32>,
-    pub script_name: Option<String>,
-    /// Total pulses to fire. `1` = single shot (most effects). `> 1` = DoT/HoT.
-    /// `0` = pulse-until-removed (channelled holdable abilities).
-    pub pulse_count: i32,
-    /// Seconds between pulses. Ignored when `pulse_count == 1`.
-    pub pulse_duration: f32,
-    /// `true` when the source must maintain channel state (cancellable
-    /// by movement, interrupt, etc.). Pulse-loop respects this; v1
-    /// treats interrupt and channel-cancel as TODO.
-    pub is_channeled: bool,
-    /// `TCM_Single` / `TCM_AERadius` / `TCM_AECone`. See module-level
-    /// `TCM_*` constants. Unknown values fall through to single-target
-    /// with a warn log so new content surfaces as a missing dispatch.
-    pub target_collection_method: String,
-    /// Range tier — one of the string tier names the original content
-    /// authoring used (Melee, Short, Medium, Long, Weapon). `Weapon`
-    /// means "use the equipped weapon's range". Pulled through to
-    /// cone/radius computation via `tcm_range_meters`.
-    pub tcm_param1: String,
-    /// Width tier for cone (Narrow, Medium, Wide); secondary range
-    /// tier for radius. Same tier-string discipline as `tcm_param1`.
-    pub tcm_param2: String,
-    /// Packed flag bitmask. See `EF_*` constants. Drives category
-    /// dispatch (Stun, Suppression, DoT, etc.).
-    pub flags: u32,
-    /// Name-value pairs: e.g. "HealthDamage" → "15"
-    pub params: std::collections::HashMap<String, String>,
-}
-
-impl Default for EffectDef {
-    /// Single-shot, single-target, no flags — the most common shape.
-    /// Test code and synthetic construction sites use this with struct
-    /// update syntax to set only the fields they care about, e.g.
-    /// `EffectDef { effect_id: 999, ability_id: 1, ..Default::default() }`.
-    fn default() -> Self {
-        Self {
-            effect_id: 0,
-            ability_id: 0,
-            delay: 0,
-            effect_sequence: 0,
-            event_set_id: None,
-            script_name: None,
-            pulse_count: 1,
-            pulse_duration: 0.0,
-            is_channeled: false,
-            target_collection_method: TCM_SINGLE.to_string(),
-            tcm_param1: String::new(),
-            tcm_param2: String::new(),
-            flags: 0,
-            params: std::collections::HashMap::new(),
-        }
-    }
-}
-
-impl EffectDef {
-    /// Get a param as i32, returning 0 if missing or unparseable.
-    pub fn param_i32(&self, name: &str) -> i32 {
-        self.params
-            .get(name)
-            .and_then(|v| v.parse().ok())
-            .unwrap_or(0)
-    }
-
-    /// Get a param as f32, returning 0.0 if missing or unparseable.
-    pub fn param_f32(&self, name: &str) -> f32 {
-        self.params
-            .get(name)
-            .and_then(|v| v.parse().ok())
-            .unwrap_or(0.0)
-    }
-
-    /// `true` when the effect should pulse over time (DoT/HoT/channelled).
-    /// Single-shot effects (`pulse_count == 1`) return `false`.
-    pub fn is_pulsing(&self) -> bool {
-        self.pulse_count == 0 || self.pulse_count > 1
-    }
-
-    /// Total duration of a pulsing effect in seconds. `0.0` for
-    /// single-shot or channelled-until-released (`pulse_count == 0`).
-    pub fn total_duration(&self) -> f32 {
-        if self.pulse_count <= 1 {
-            0.0
-        } else {
-            self.pulse_count as f32 * self.pulse_duration.max(0.0)
-        }
-    }
-
-    /// Translate a TCM range/width tier name to meters. The fan-server's
-    /// tier names come from the original 2009 content authoring — we map
-    /// them to concrete distances tuned against the radius defaults
-    /// already used elsewhere in combat (e.g. `DEFAULT_GROUND_TARGET_RADIUS = 5.0`).
-    ///
-    /// Unknown tiers fall back to `Medium = 8.0m` with a warn so new
-    /// content surfaces as a tier we need to calibrate.
-    pub fn tcm_range_meters(tier: &str) -> f32 {
-        match tier {
-            "Melee" => 2.5,
-            "Short" => 5.0,
-            "Medium" => 8.0,
-            "Long" => 14.0,
-            "Weapon" => 20.0, // matches the canonical weapon-range cap
-            other => {
-                tracing::warn!(
-                    target: "abilities",
-                    event = "tcm_unknown_range_tier",
-                    tier = other,
-                    "Unknown TCM range tier; defaulting to Medium=8.0m"
-                );
-                8.0
-            }
-        }
-    }
-
-    /// Translate a TCM width tier to a cone half-angle in radians.
-    /// Narrow / Medium / Wide are the only width tiers in seed.
-    pub fn tcm_half_angle_radians(tier: &str) -> f32 {
-        // Half-angles chosen so:
-        //   Narrow = 22.5° half-angle (45° full cone) — shotgun spread
-        //   Medium = 45°               (90° full)     — grenade arc
-        //   Wide   = 67.5°             (135° full)    — flamethrower-class
-        match tier {
-            "Narrow" => std::f32::consts::FRAC_PI_8,     // 22.5° = π/8
-            "Medium" => std::f32::consts::FRAC_PI_4,     // 45° = π/4
-            "Wide" => 3.0 * std::f32::consts::FRAC_PI_8, // 67.5° = 3π/8
-            other => {
-                tracing::warn!(
-                    target: "abilities",
-                    event = "tcm_unknown_width_tier",
-                    tier = other,
-                    "Unknown TCM width tier; defaulting to Medium=45°"
-                );
-                std::f32::consts::FRAC_PI_4
-            }
-        }
-    }
-}
-
-// ── AbilityTreeData ───────────────────────────────────────────────────────
-
-/// Archetype ability tree: 3 branches of ability IDs.
-///
-/// Sent to the client via `onAbilityTreeInfo(ARRAY<ARRAY<INT32>>)`.
-/// Source: `resources.archetype_ability_tree` table, grouped by `tree_index` (0-2).
-#[derive(Debug, Clone, Default)]
-pub struct AbilityTreeData {
-    pub trees: [Vec<i32>; 3],
-}
-
-impl AbilityTreeData {
-    /// Serialize for `onAbilityTreeInfo(ARRAY<ARRAY<INT32>>)`.
-    ///
-    /// The .def declares a single nested-array `<Arg>` —
-    /// `ARRAY <of> ARRAY <of> INT32 </of> </of>` — so BigWorld's wire
-    /// encoding is `[outer_count:u32][inner_count:u32 ids...] × N`. Both
-    /// counts are required; the outer count is what tells the client how
-    /// many inner arrays to expect. Without it the client reads the first
-    /// inner count as the outer count, tries to parse that many inner
-    /// arrays, the parser fails, and the handler aborts silently — the
-    /// abilities window stays empty.
-    ///
-    /// SGW's authored content always sends exactly 3 trees, but the
-    /// outer count is on the wire either way. The Python
-    /// `self.client.onAbilityTreeInfo([a, b, c])` call passes a 3-element
-    /// list; the BigWorld serializer on the Python client side still
-    /// prepends the outer length prefix.
-    ///
-    /// Compare with the working `setupStargateInfo` shape: that .def has
-    /// THREE separate `<Arg>` entries, each `ARRAY<of>INT32</of>`. Three
-    /// top-level args, each its own length-prefixed array, no outer
-    /// wrapper. Coincidentally the same byte layout up to the outer
-    /// prefix — which is what masked this bug.
-    pub fn serialize(&self) -> Vec<u8> {
-        let total_ids: usize = self.trees.iter().map(|t| t.len()).sum();
-        let mut buf = Vec::with_capacity(4 + 4 * self.trees.len() + total_ids * 4);
-        buf.extend_from_slice(&(self.trees.len() as u32).to_le_bytes());
-        for tree in &self.trees {
-            buf.extend_from_slice(&(tree.len() as u32).to_le_bytes());
-            for &id in tree {
-                buf.extend_from_slice(&id.to_le_bytes());
-            }
-        }
-        buf
-    }
-}
-
-// ── Cooldown tracking ─────────────────────────────────────────────────────
-
-/// A cooldown entry for an ability or moniker group.
-#[derive(Debug, Clone)]
-pub struct CooldownEntry {
-    /// When the cooldown expires.
-    pub expires_at: Instant,
-    /// Total cooldown duration (for client timer display).
-    pub total_duration: Duration,
-}
-
-// ── AbilityManager ────────────────────────────────────────────────────────
-
-/// Per-entity ability manager.
-///
-/// Tracks known abilities, cooldowns, and auto-cycle state.
-/// Mirrors `python/cell/AbilityManager.py`.
-#[derive(Debug)]
-pub struct AbilityManager {
-    /// Set of ability IDs this entity knows.
-    known_abilities: HashSet<i32>,
-
-    /// Subset of [`Self::known_abilities`] that were added on behalf of the
-    /// **currently-equipped weapon** via the `items_event_sets` binding
-    /// (e.g., equipping a P90 grants the SMG Auto Attack ability 559).
-    ///
-    /// Maintained separately from `known_abilities` so weapon-equip
-    /// swaps can revoke exactly the abilities the *previous* weapon
-    /// added, without touching player-trained or archetype-starter
-    /// abilities. An ability is tracked here ONLY when the weapon
-    /// binding added it — if the player already knew the same id
-    /// (trained it, archetype-granted), we leave it untracked and
-    /// never revoke it on weapon swap.
-    ///
-    /// See [`Self::swap_weapon_granted_abilities`] for the diff math.
-    ///
-    /// **Runtime-only, not persisted.** The set is rebuilt on
-    /// session start: when a player re-enters the world, the
-    /// bandolier's `active_bandolier_slot` is loaded from the
-    /// `bandolier_slot` column, and the cell's
-    /// `swap_weapon_granted_abilities_for_slot` re-runs against an
-    /// empty tracked set + the loaded weapon — producing the same
-    /// granted set the previous session ended with. The DB doesn't
-    /// need a column for this; the items_event_sets table + the
-    /// active slot are the canonical source.
-    weapon_granted_abilities: HashSet<i32>,
-
-    /// Active ability cooldowns keyed by ability_id.
-    ability_cooldowns: HashMap<i32, CooldownEntry>,
-
-    /// Active moniker cooldowns keyed by moniker_id.
-    moniker_cooldowns: HashMap<i64, CooldownEntry>,
-
-    /// Whether auto-cycle is enabled.
-    ///
-    /// Set by the `setAutoCycle(enabled=1)` cell method (gun-icon button) and
-    /// cleared by `setAutoCycle(0)`, by `stoppedAutoCycling`-equivalent events
-    /// (target death, `AF_DEACTIVATE_AUTO_CYCLE` ability fire), or by a manual
-    /// `useAbility` against a different ability ID. The actual re-fire loop
-    /// lives in `cimmeria-services` (`cell::service::ticks::auto_cycle_tick`) —
-    /// this flag just arms it.
-    pub auto_cycle: bool,
-
-    /// The ability ID being auto-cycled (if any).
-    ///
-    /// Stashed at the first commit of `useAbility` after `auto_cycle` flipped
-    /// to `true`. The driver tick re-invokes `handle_use_ability` with this
-    /// id every time the cooldown clears.
-    pub auto_cycle_ability_id: Option<i32>,
-
-    /// The last ability the player successfully fired. Stashed at every
-    /// `handle_use_ability` commit regardless of `auto_cycle` state, so
-    /// `setAutoCycle(1)` can fire immediately against the player's
-    /// previously-fired ability + current target without requiring the
-    /// player to right-click an enemy first.
-    ///
-    /// `None` only at session start before any fire. Once set, kept
-    /// for the rest of the session — never cleared on auto-cycle
-    /// stop, death, or respawn. Stale values are harmless: the
-    /// immediate-fire path runs the value through the normal
-    /// `handle_use_ability` validation, which rejects abilities the
-    /// player no longer has (e.g. after a weapon unequip) and leaves
-    /// the loop BSF-armed for the next manual fire to refresh.
-    ///
-    /// Distinct from `auto_cycle_ability_id`: that field is the ability
-    /// the LOOP is currently committed to and clears on stop; this
-    /// field is the player's most-recent fire, used as the seed for
-    /// the immediate-fire-on-enable path.
-    pub last_fired_ability_id: Option<i32>,
-
-    /// Next unique effect sequence ID.
-    pub effect_sequence_id: i32,
-}
-
-impl AbilityManager {
-    /// Create a new empty ability manager.
-    pub fn new() -> Self {
-        Self {
-            known_abilities: HashSet::new(),
-            weapon_granted_abilities: HashSet::new(),
-            ability_cooldowns: HashMap::new(),
-            moniker_cooldowns: HashMap::new(),
-            auto_cycle: false,
-            auto_cycle_ability_id: None,
-            last_fired_ability_id: None,
-            effect_sequence_id: 1,
-        }
-    }
-
-    /// Create a new ability manager with the given known ability IDs.
-    pub fn with_abilities(ability_ids: &[i32]) -> Self {
-        let mut mgr = Self::new();
-        for &id in ability_ids {
-            mgr.known_abilities.insert(id);
-        }
-        mgr
-    }
-
-    /// Check if the entity knows a given ability.
-    pub fn has_ability(&self, ability_id: i32) -> bool {
-        self.known_abilities.contains(&ability_id)
-    }
-
-    /// Add a known ability.
-    pub fn add_ability(&mut self, ability_id: i32) {
-        self.known_abilities.insert(ability_id);
-    }
-
-    /// Get the first known ability (for NPCs using their default attack).
-    pub fn first_known_ability(&self) -> Option<i32> {
-        self.known_abilities.iter().next().copied()
-    }
-
-    /// Remove a known ability.
-    pub fn remove_ability(&mut self, ability_id: i32) {
-        self.known_abilities.remove(&ability_id);
-        // If the ability was tracked as weapon-granted, drop the tracking
-        // too — caller is removing it unconditionally, so the
-        // weapon-grant accounting shouldn't outlive the underlying
-        // known-set entry.
-        self.weapon_granted_abilities.remove(&ability_id);
-    }
-
-    /// Swap the set of abilities granted by the currently-equipped
-    /// weapon for a new set (i.e., player equipped a different weapon).
-    ///
-    /// Returns `(removed, added)` — the ability IDs that left and
-    /// joined `known_abilities` as a result of this call. An empty
-    /// `(removed, added)` means nothing changed and the caller should
-    /// skip the `onKnownAbilitiesUpdate` broadcast to avoid wire noise.
-    ///
-    /// Semantics:
-    ///
-    /// 1. For each id in the **previous** `weapon_granted_abilities`
-    ///    that is NOT in the new set: remove it from `known_abilities`
-    ///    AND drop the weapon-grant tag. This is the revoke path —
-    ///    the old weapon no longer grants it and we added it on the
-    ///    weapon's behalf, so it leaves with the weapon.
-    /// 2. For each id in the **new** set that the player does NOT
-    ///    already know: add it to `known_abilities` AND tag it as
-    ///    weapon-granted. The tag is what makes step 1 work on the
-    ///    NEXT swap.
-    /// 3. For each id in the **new** set that the player already
-    ///    knows (trained, archetype-granted, etc.): leave
-    ///    `known_abilities` untouched AND do NOT tag it as weapon-
-    ///    granted. The player owns this ability independently of the
-    ///    weapon and we won't revoke it on unequip.
-    ///
-    /// Passing an empty `new_set` is the "unequip — no weapon" call:
-    /// every previously-tracked weapon ability is revoked, and nothing
-    /// is added.
-    ///
-    /// **Why a separate tracked set rather than computing the diff
-    /// from items_event_sets at unequip time?** The previous weapon
-    /// may have been removed from the items table mid-session
-    /// (content reload, hot-fix), or its bindings may have changed.
-    /// Tracking what we actually added means we can always undo
-    /// exactly that, independent of the current state of the data.
-    pub fn swap_weapon_granted_abilities(&mut self, new_set: HashSet<i32>) -> (Vec<i32>, Vec<i32>) {
-        // Removed: previously weapon-granted but not in the new set.
-        let removed: Vec<i32> = self
-            .weapon_granted_abilities
-            .difference(&new_set)
-            .copied()
-            .collect();
-        for id in &removed {
-            self.known_abilities.remove(id);
-            self.weapon_granted_abilities.remove(id);
-        }
-
-        // Added: new entries the player didn't already know. If they
-        // already knew it (trained/archetype), the weapon-grant tag
-        // isn't taken — that ability stays player-owned and won't be
-        // revoked on the next swap.
-        let mut added: Vec<i32> = Vec::new();
-        for id in new_set {
-            if !self.known_abilities.contains(&id) {
-                self.known_abilities.insert(id);
-                self.weapon_granted_abilities.insert(id);
-                added.push(id);
-            }
-        }
-
-        (removed, added)
-    }
-
-    /// Inspect which abilities are currently tagged as weapon-granted.
-    /// Test helper; exposed for verification, not for runtime mutation.
-    #[doc(hidden)]
-    pub fn weapon_granted_ability_ids(&self) -> Vec<i32> {
-        self.weapon_granted_abilities.iter().copied().collect()
-    }
-
-    /// Get all known ability IDs (unsorted).
-    pub fn known_ability_ids(&self) -> Vec<i32> {
-        self.known_abilities.iter().copied().collect()
-    }
-
-    /// Number of known abilities.
-    pub fn known_count(&self) -> usize {
-        self.known_abilities.len()
-    }
-
-    /// Check if an ability is on cooldown.
-    pub fn is_on_cooldown(&self, ability_id: i32) -> bool {
-        if let Some(entry) = self.ability_cooldowns.get(&ability_id) {
-            Instant::now() < entry.expires_at
-        } else {
-            false
-        }
-    }
-
-    /// Check if a moniker group is on cooldown.
-    pub fn is_moniker_on_cooldown(&self, moniker_id: i64) -> bool {
-        if let Some(entry) = self.moniker_cooldowns.get(&moniker_id) {
-            Instant::now() < entry.expires_at
-        } else {
-            false
-        }
-    }
-
-    /// Start a cooldown for an ability.
-    pub fn start_ability_cooldown(&mut self, ability_id: i32, duration: Duration) {
-        self.ability_cooldowns.insert(
-            ability_id,
-            CooldownEntry {
-                expires_at: Instant::now() + duration,
-                total_duration: duration,
-            },
-        );
-    }
-
-    /// Start cooldowns for an ability and its moniker groups.
-    pub fn start_cooldowns(&mut self, ability: &AbilityDef, cooldown_secs: f32) {
-        let duration = Duration::from_secs_f32(cooldown_secs);
-        self.start_ability_cooldown(ability.ability_id, duration);
-
-        // Moniker cooldowns expire slightly before ability cooldown
-        // to prevent race conditions (Python: completeTime - 0.001)
-        let moniker_duration = if cooldown_secs > 0.002 {
-            Duration::from_secs_f32(cooldown_secs - 0.001)
-        } else {
-            duration
-        };
-        for &moniker_id in &ability.moniker_ids {
-            self.moniker_cooldowns.insert(
-                moniker_id,
-                CooldownEntry {
-                    expires_at: Instant::now() + moniker_duration,
-                    total_duration: moniker_duration,
-                },
-            );
-        }
-    }
-
-    /// Clear all ability and moniker cooldowns (e.g., on respawn).
-    pub fn clear_all_cooldowns(&mut self) {
-        self.ability_cooldowns.clear();
-        self.moniker_cooldowns.clear();
-    }
-
-    /// Remove expired cooldowns. Call periodically (e.g., each tick).
-    pub fn cleanup_expired(&mut self) {
-        let now = Instant::now();
-        self.ability_cooldowns.retain(|_, e| now < e.expires_at);
-        self.moniker_cooldowns.retain(|_, e| now < e.expires_at);
-    }
-
-    /// Check if an ability can be used (basic validation).
-    ///
-    /// Returns `Ok(())` or an error string describing why not.
-    pub fn can_use_ability(&self, ability: &AbilityDef) -> Result<(), &'static str> {
-        if !self.has_ability(ability.ability_id) {
-            return Err("entity does not have ability");
-        }
-        if self.is_on_cooldown(ability.ability_id) {
-            return Err("ability on cooldown");
-        }
-        for &moniker_id in &ability.moniker_ids {
-            if self.is_moniker_on_cooldown(moniker_id) {
-                return Err("moniker on cooldown");
-            }
-        }
-        Ok(())
-    }
-
-    /// Get next unique effect sequence ID and increment.
-    pub fn next_effect_id(&mut self) -> i32 {
-        let id = self.effect_sequence_id;
-        self.effect_sequence_id += 1;
-        id
-    }
-
-    /// Serialize known abilities for `onKnownAbilitiesUpdate(ARRAY<INT32>)`.
-    ///
-    /// Wire: `count:u32 [ability_id:i32 ...]`.
-    pub fn serialize_known_abilities(&self) -> Vec<u8> {
-        let ids: Vec<i32> = self.known_abilities.iter().copied().collect();
-        let mut buf = Vec::with_capacity(4 + ids.len() * 4);
-        buf.extend_from_slice(&(ids.len() as u32).to_le_bytes());
-        for id in &ids {
-            buf.extend_from_slice(&id.to_le_bytes());
-        }
-        buf
-    }
-}
-
-impl Default for AbilityManager {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-// ── Wire format helpers ───────────────────────────────────────────────────
-
-/// A single stat change result sent to the client in `onEffectResults`.
-///
-/// Wire per entry: `stat_id:i8, delta:i32, damage_code:i8, stat_result_code:i8` (7 bytes).
-#[derive(Debug, Clone)]
-pub struct ClientEffectResult {
-    pub stat_id: i8,
-    pub delta: i32,
-    pub damage_code: i8,
-    pub stat_result_code: i8,
-}
-
-impl ClientEffectResult {
-    pub fn serialize(&self, buf: &mut Vec<u8>) {
-        buf.push(self.stat_id as u8);
-        buf.extend_from_slice(&self.delta.to_le_bytes());
-        buf.push(self.damage_code as u8);
-        buf.push(self.stat_result_code as u8);
-    }
-}
-
-/// Serialize `onTimerUpdate` arguments.
-///
-/// Wire: `id:i32, type:i8, sourceId:i32, secondaryId:i32, totalTime:f32, expireTime:f32`.
-pub fn serialize_timer_update(
-    id: i32,
-    timer_type: i8,
-    source_id: i32,
-    total_time: f32,
-    expire_time: f32,
-) -> Vec<u8> {
-    let mut buf = Vec::with_capacity(21);
-    buf.extend_from_slice(&id.to_le_bytes());
-    buf.push(timer_type as u8);
-    buf.extend_from_slice(&source_id.to_le_bytes());
-    buf.extend_from_slice(&0i32.to_le_bytes()); // secondaryId always 0
-    buf.extend_from_slice(&total_time.to_le_bytes());
-    buf.extend_from_slice(&expire_time.to_le_bytes());
-    buf
-}
-
-/// Serialize `onEffectResults` arguments.
-///
-/// Wire: `sourceId:i32, abilityId:i32, effectId:i32, targetId:i32,
-///        resultCode:u8, count:u32, [ClientEffectResult...]`.
-pub fn serialize_effect_results(
-    source_id: i32,
-    ability_id: i32,
-    effect_id: i32,
-    target_id: i32,
-    result_code: u8,
-    stat_results: &[ClientEffectResult],
-) -> Vec<u8> {
-    let mut buf = Vec::with_capacity(21 + stat_results.len() * 7);
-    buf.extend_from_slice(&source_id.to_le_bytes());
-    buf.extend_from_slice(&ability_id.to_le_bytes());
-    buf.extend_from_slice(&effect_id.to_le_bytes());
-    buf.extend_from_slice(&target_id.to_le_bytes());
-    buf.push(result_code);
-    buf.extend_from_slice(&(stat_results.len() as u32).to_le_bytes());
-    for result in stat_results {
-        result.serialize(&mut buf);
-    }
-    buf
-}
-
-// ── Tests ─────────────────────────────────────────────────────────────────
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    fn test_ability() -> AbilityDef {
-        AbilityDef {
-            ability_id: 597,
-            name: "Rapid Fire".to_string(),
-            cooldown: 5.0,
-            warmup: 1.5,
-            flags: AF_SPEED_ATTACK,
-            is_ranged: true,
-            min_range: 0,
-            max_range: 30,
-            target_type_id: TARGET_TARGET,
-            effect_ids: vec![100, 101],
-            moniker_ids: vec![42],
-            required_ammo: 1,
-            event_set_id: None,
-            velocity: 0.0,
-        }
-    }
-
-    #[test]
-    fn new_manager_is_empty() {
-        let mgr = AbilityManager::new();
-        assert_eq!(mgr.known_count(), 0);
-        assert!(!mgr.auto_cycle);
-        assert_eq!(mgr.effect_sequence_id, 1);
-    }
-
-    #[test]
-    fn add_and_check_abilities() {
-        let mut mgr = AbilityManager::new();
-        mgr.add_ability(597);
-        mgr.add_ability(592);
-        assert!(mgr.has_ability(597));
-        assert!(mgr.has_ability(592));
-        assert!(!mgr.has_ability(999));
-        assert_eq!(mgr.known_count(), 2);
-    }
-
-    #[test]
-    fn remove_ability() {
-        let mut mgr = AbilityManager::with_abilities(&[597, 592, 594]);
-        assert_eq!(mgr.known_count(), 3);
-        mgr.remove_ability(592);
-        assert!(!mgr.has_ability(592));
-        assert_eq!(mgr.known_count(), 2);
-    }
-
-    #[test]
-    fn with_abilities_constructor() {
-        let mgr = AbilityManager::with_abilities(&[1, 2, 3]);
-        assert_eq!(mgr.known_count(), 3);
-        assert!(mgr.has_ability(1));
-        assert!(mgr.has_ability(2));
-        assert!(mgr.has_ability(3));
-    }
-
-    #[test]
-    fn cooldown_tracking() {
-        let mut mgr = AbilityManager::with_abilities(&[597]);
-        let ability = test_ability();
-
-        assert!(!mgr.is_on_cooldown(597));
-        mgr.start_ability_cooldown(597, Duration::from_secs(5));
-        assert!(mgr.is_on_cooldown(597));
-
-        // Moniker cooldown
-        assert!(!mgr.is_moniker_on_cooldown(42));
-        mgr.start_cooldowns(&ability, 5.0);
-        assert!(mgr.is_moniker_on_cooldown(42));
-    }
-
-    #[test]
-    fn can_use_ability_checks() {
-        let mgr = AbilityManager::with_abilities(&[597]);
-        let ability = test_ability();
-        assert!(mgr.can_use_ability(&ability).is_ok());
-
-        // Unknown ability
-        let unknown = AbilityDef {
-            ability_id: 999,
-            ..test_ability()
-        };
-        assert_eq!(
-            mgr.can_use_ability(&unknown).unwrap_err(),
-            "entity does not have ability"
-        );
-    }
-
-    #[test]
-    fn can_use_ability_cooldown_check() {
-        let mut mgr = AbilityManager::with_abilities(&[597]);
-        let ability = test_ability();
-        mgr.start_ability_cooldown(597, Duration::from_secs(5));
-        assert_eq!(
-            mgr.can_use_ability(&ability).unwrap_err(),
-            "ability on cooldown"
-        );
-    }
-
-    #[test]
-    fn next_effect_id_increments() {
-        let mut mgr = AbilityManager::new();
-        assert_eq!(mgr.next_effect_id(), 1);
-        assert_eq!(mgr.next_effect_id(), 2);
-        assert_eq!(mgr.next_effect_id(), 3);
-    }
-
-    #[test]
-    fn serialize_known_abilities() {
-        let mgr = AbilityManager::with_abilities(&[597, 592]);
-        let data = mgr.serialize_known_abilities();
-        let count = u32::from_le_bytes([data[0], data[1], data[2], data[3]]);
-        assert_eq!(count, 2);
-        assert_eq!(data.len(), 4 + 2 * 4); // count + 2 ids
-    }
-
-    // ── swap_weapon_granted_abilities ────────────────────────────────
-
-    /// First weapon equip — empty tracked set, non-empty new set:
-    /// every new id lands in known_abilities AND in
-    /// weapon_granted_abilities. `added` returns the full new set;
-    /// `removed` is empty.
-    #[test]
-    fn swap_weapon_granted_first_equip_adds_all_new() {
-        let mut mgr = AbilityManager::with_abilities(&[594, 597]);
-        // Pre: archetype starter has Strike (594) + Heal Focus (597).
-        // Equip P90 — items_event_sets binds 559 (SMG Auto) + 595 (SMG Melee).
-        let new_set: HashSet<i32> = [559, 595].into_iter().collect();
-        let (removed, added) = mgr.swap_weapon_granted_abilities(new_set);
-
-        assert!(removed.is_empty(), "first equip removes nothing");
-        let mut added_sorted = added.clone();
-        added_sorted.sort();
-        assert_eq!(added_sorted, vec![559, 595]);
-        assert!(mgr.has_ability(559));
-        assert!(mgr.has_ability(595));
-        // Archetype abilities untouched.
-        assert!(mgr.has_ability(594));
-        assert!(mgr.has_ability(597));
-        // Weapon-grant tags hold exactly the added ids.
-        let mut weapon_granted = mgr.weapon_granted_ability_ids();
-        weapon_granted.sort();
-        assert_eq!(weapon_granted, vec![559, 595]);
-    }
-
-    /// Swap weapon — old tracked abilities revoke, new ones grant.
-    /// Player's archetype abilities are NOT in either diff.
-    #[test]
-    fn swap_weapon_granted_weapon_swap_revokes_old_grants_new() {
-        let mut mgr = AbilityManager::with_abilities(&[594, 597]);
-        // Equip pistol first — adds 579 + 708 to known + tracked.
-        let pistol_set: HashSet<i32> = [579, 708].into_iter().collect();
-        mgr.swap_weapon_granted_abilities(pistol_set);
-
-        // Then swap to P90 — adds 559 + 595, revokes 579 + 708.
-        let p90_set: HashSet<i32> = [559, 595].into_iter().collect();
-        let (removed, added) = mgr.swap_weapon_granted_abilities(p90_set);
-
-        let mut removed_sorted = removed.clone();
-        removed_sorted.sort();
-        assert_eq!(removed_sorted, vec![579, 708]);
-
-        let mut added_sorted = added.clone();
-        added_sorted.sort();
-        assert_eq!(added_sorted, vec![559, 595]);
-
-        // Pistol abilities gone, P90 abilities present, archetype intact.
-        assert!(!mgr.has_ability(579));
-        assert!(!mgr.has_ability(708));
-        assert!(mgr.has_ability(559));
-        assert!(mgr.has_ability(595));
-        assert!(mgr.has_ability(594));
-        assert!(mgr.has_ability(597));
-    }
-
-    /// Unequip — empty new set revokes every tracked weapon ability
-    /// and leaves the player with just their non-weapon known set.
-    #[test]
-    fn swap_weapon_granted_unequip_revokes_all_tracked() {
-        let mut mgr = AbilityManager::with_abilities(&[594, 597]);
-        mgr.swap_weapon_granted_abilities([579, 708].into_iter().collect());
-
-        let (removed, added) = mgr.swap_weapon_granted_abilities(HashSet::new());
-
-        let mut removed_sorted = removed.clone();
-        removed_sorted.sort();
-        assert_eq!(removed_sorted, vec![579, 708]);
-        assert!(added.is_empty());
-        assert!(!mgr.has_ability(579));
-        assert!(!mgr.has_ability(708));
-        assert!(mgr.has_ability(594));
-        assert!(mgr.has_ability(597));
-        assert!(mgr.weapon_granted_ability_ids().is_empty());
-    }
-
-    /// **The player-already-knew-it carve-out.** If the new weapon's
-    /// binding overlaps an ability the player ALREADY had (trained,
-    /// archetype-granted, or already weapon-granted), the swap MUST
-    /// NOT tag it as weapon-granted — otherwise the next unequip
-    /// would revoke a player-owned ability.
-    ///
-    /// Reverting the `!self.known_abilities.contains(&id)` gate around
-    /// the weapon-grant insert trips this guard by re-tagging an
-    /// already-known ability, which then gets revoked on the next
-    /// swap — the assertion `has_ability(594)` fails after the second
-    /// swap.
-    #[test]
-    fn swap_weapon_granted_skips_tagging_already_known_ability() {
-        // Player has Strike (594) trained as part of archetype.
-        let mut mgr = AbilityManager::with_abilities(&[594]);
-        // Equip a hypothetical weapon that ALSO binds 594 + 559.
-        let weapon_set: HashSet<i32> = [594, 559].into_iter().collect();
-        let (_removed, added) = mgr.swap_weapon_granted_abilities(weapon_set);
-
-        // 594 was already known — must NOT appear in `added` (it
-        // didn't transition into the known set) and must NOT be
-        // tagged.
-        assert_eq!(added, vec![559], "only 559 transitioned into known");
-        let weapon_granted = mgr.weapon_granted_ability_ids();
-        assert_eq!(
-            weapon_granted,
-            vec![559],
-            "594 must NOT be tagged as weapon-granted — player already knew it"
-        );
-
-        // Now swap to a weapon that grants neither 594 nor 559.
-        // 559 (was tracked) gets revoked; 594 STAYS because it was
-        // never tagged.
-        let other_set: HashSet<i32> = [777].into_iter().collect();
-        let (removed2, _added2) = mgr.swap_weapon_granted_abilities(other_set);
-        assert_eq!(removed2, vec![559]);
-        assert!(
-            mgr.has_ability(594),
-            "player-owned 594 must survive the second swap — \
-             reverting the already-known carve-out makes this fail"
-        );
-    }
-
-    /// No-op swap — old and new sets are identical → empty diff,
-    /// no broadcast warranted. Pins the empty-tuple short-circuit
-    /// the bandolier handler keys on.
-    #[test]
-    fn swap_weapon_granted_identical_set_returns_empty_diff() {
-        let mut mgr = AbilityManager::with_abilities(&[594, 597]);
-        mgr.swap_weapon_granted_abilities([559, 595].into_iter().collect());
-
-        // Re-apply the same set.
-        let (removed, added) = mgr.swap_weapon_granted_abilities([559, 595].into_iter().collect());
-        assert!(removed.is_empty());
-        assert!(added.is_empty());
-    }
-
-    #[test]
-    fn ability_tree_empty() {
-        let tree = AbilityTreeData::default();
-        let data = tree.serialize();
-        // outer count(4) + 3 inner counts(4) = 16
-        assert_eq!(data.len(), 16);
-        // Outer count = 3 (number of inner arrays)
-        let outer = u32::from_le_bytes([data[0], data[1], data[2], data[3]]);
-        assert_eq!(outer, 3);
-        // Each of the 3 inner counts = 0
-        for i in 0..3 {
-            let off = 4 + i * 4;
-            let count =
-                u32::from_le_bytes([data[off], data[off + 1], data[off + 2], data[off + 3]]);
-            assert_eq!(count, 0, "inner array {i} count must be 0");
-        }
-    }
-
-    #[test]
-    fn ability_tree_with_data() {
-        let tree = AbilityTreeData {
-            trees: [vec![592, 1005, 642], vec![597, 646, 641], vec![]],
-        };
-        let data = tree.serialize();
-        // outer(4) + tree0:count(4)+3*4=12=16 + tree1:16 + tree2:count(4)=4 = 40
-        assert_eq!(data.len(), 40);
-
-        // Outer count = 3 inner arrays
-        let outer = u32::from_le_bytes([data[0], data[1], data[2], data[3]]);
-        assert_eq!(outer, 3);
-        // First inner count = 3 ids
-        let c0 = u32::from_le_bytes([data[4], data[5], data[6], data[7]]);
-        assert_eq!(c0, 3);
-        // First ability id in tree 0 = 592
-        let a0 = i32::from_le_bytes([data[8], data[9], data[10], data[11]]);
-        assert_eq!(a0, 592);
-    }
-
-    /// **Regression guard: the outer ARRAY count prefix is on the wire.**
-    ///
-    /// Bug shape: the wire encoding for `ARRAY<ARRAY<INT32>>` is
-    /// `[outer:u32][inner:u32 ids...] × N`. Skipping the outer prefix
-    /// is what the original Rust impl did — based on the (wrong) theory
-    /// that "the Python code always sends exactly 3 arrays so the count
-    /// is implicit." It isn't; BigWorld's serializer still emits it.
-    /// The client decoder reads the first u32 as the outer count, gets
-    /// some inner-tree size like 29, tries to parse 29 inner arrays,
-    /// parser fails, handler aborts silently → abilities window stays
-    /// empty for every player.
-    ///
-    /// Reverting [`AbilityTreeData::serialize`] to skip the outer count
-    /// must fail this assertion.
-    #[test]
-    fn ability_tree_outer_count_prefix_is_first_four_bytes() {
-        let tree = AbilityTreeData {
-            trees: [
-                // Real shape: Soldier tree 0 has 29 abilities; if we read
-                // the first u32 as the outer count we'd get 29, not 3 —
-                // the exact decode error the client makes on unprefixed
-                // bytes.
-                (0..29).collect(),
-                vec![],
-                vec![],
-            ],
-        };
-        let data = tree.serialize();
-        let outer = u32::from_le_bytes([data[0], data[1], data[2], data[3]]);
-        assert_eq!(
-            outer, 3,
-            "first 4 bytes MUST be the outer ARRAY count (=3, the number of \
-             trees). Reverting serialize() to skip the outer prefix would \
-             surface this as outer=29 here — the exact byte the client \
-             misreads as outer-count on the wire."
-        );
-    }
-
-    #[test]
-    fn serialize_timer_update_format() {
-        let data = serialize_timer_update(597, TIMER_ABILITY_COOLDOWN, 100, 5.0, 12345.0);
-        assert_eq!(data.len(), 21);
-        let id = i32::from_le_bytes([data[0], data[1], data[2], data[3]]);
-        assert_eq!(id, 597);
-        assert_eq!(data[4], TIMER_ABILITY_COOLDOWN as u8);
-    }
-
-    #[test]
-    fn serialize_effect_results_empty() {
-        let data = serialize_effect_results(100, 597, 101, 200, RC_HIT, &[]);
-        assert_eq!(data.len(), 21); // 4*4 + 1 + 4 = 21
-        assert_eq!(data[16], RC_HIT);
-        let count = u32::from_le_bytes([data[17], data[18], data[19], data[20]]);
-        assert_eq!(count, 0);
-    }
-
-    #[test]
-    fn serialize_effect_results_with_stats() {
-        let results = vec![ClientEffectResult {
-            stat_id: 10, // health
-            delta: -50,
-            damage_code: DT_PHYSICAL,
-            stat_result_code: SRC_NONE,
-        }];
-        let data = serialize_effect_results(100, 597, 101, 200, RC_CRITICAL, &results);
-        assert_eq!(data.len(), 21 + 7); // 21 header + 7 per result
-        assert_eq!(data[16], RC_CRITICAL);
-        let count = u32::from_le_bytes([data[17], data[18], data[19], data[20]]);
-        assert_eq!(count, 1);
-        // stat_id
-        assert_eq!(data[21], 10);
-    }
-
-    #[test]
-    fn client_effect_result_serialize() {
-        let r = ClientEffectResult {
-            stat_id: 10,
-            delta: -100,
-            damage_code: DT_ENERGY,
-            stat_result_code: SRC_ABSORB,
-        };
-        let mut buf = Vec::new();
-        r.serialize(&mut buf);
-        assert_eq!(buf.len(), 7);
-        assert_eq!(buf[0], 10); // stat_id
-        let delta = i32::from_le_bytes([buf[1], buf[2], buf[3], buf[4]]);
-        assert_eq!(delta, -100);
-        assert_eq!(buf[5], DT_ENERGY as u8);
-        assert_eq!(buf[6], SRC_ABSORB as u8);
-    }
-
-    #[test]
-    fn cleanup_expired_removes_old() {
-        let mut mgr = AbilityManager::with_abilities(&[597]);
-        // Add a cooldown that already expired
-        mgr.ability_cooldowns.insert(
-            597,
-            CooldownEntry {
-                expires_at: Instant::now() - Duration::from_secs(1),
-                total_duration: Duration::from_secs(5),
-            },
-        );
-        assert_eq!(mgr.ability_cooldowns.len(), 1);
-        mgr.cleanup_expired();
-        assert_eq!(mgr.ability_cooldowns.len(), 0);
-    }
-
-    // ── first_known_ability ──────────────────────────────────────────────
-
-    #[test]
-    fn first_known_ability_returns_some() {
-        let mgr = AbilityManager::with_abilities(&[597, 600, 605]);
-        let first = mgr.first_known_ability();
-        assert!(first.is_some());
-        // HashSet iteration order is non-deterministic, but it must be one of the three
-        assert!([597, 600, 605].contains(&first.unwrap()));
-    }
-
-    #[test]
-    fn first_known_ability_returns_none_when_empty() {
-        let mgr = AbilityManager::new();
-        assert!(mgr.first_known_ability().is_none());
-    }
-
-    #[test]
-    fn first_known_ability_single_element() {
-        let mut mgr = AbilityManager::new();
-        mgr.add_ability(597);
-        assert_eq!(mgr.first_known_ability(), Some(597));
-    }
-
-    // ── EffectDef helpers ────────────────────────────────────────────────
-
-    #[test]
-    fn effect_def_default_is_single_target_single_pulse() {
-        let e = EffectDef::default();
-        assert_eq!(e.pulse_count, 1, "default pulse_count=1");
-        assert_eq!(e.target_collection_method, TCM_SINGLE);
-        assert!(!e.is_pulsing(), "single-pulse is not pulsing");
-    }
-
-    #[test]
-    fn effect_def_is_pulsing_matches_pulse_count() {
-        // Channelled (pulse_count=0): is_pulsing → true
-        let channelled = EffectDef {
-            pulse_count: 0,
-            ..Default::default()
-        };
-        assert!(channelled.is_pulsing());
-        // Single shot (pulse_count=1): is_pulsing → false
-        let single = EffectDef {
-            pulse_count: 1,
-            ..Default::default()
-        };
-        assert!(!single.is_pulsing());
-        // DoT (pulse_count>1): is_pulsing → true
-        let dot = EffectDef {
-            pulse_count: 5,
-            ..Default::default()
-        };
-        assert!(dot.is_pulsing());
-    }
-
-    #[test]
-    fn effect_def_total_duration_skips_single_shot_and_channelled() {
-        let single = EffectDef {
-            pulse_count: 1,
-            pulse_duration: 5.0,
-            ..Default::default()
-        };
-        assert_eq!(single.total_duration(), 0.0, "single-shot has 0 duration");
-
-        let channelled = EffectDef {
-            pulse_count: 0,
-            pulse_duration: 5.0,
-            ..Default::default()
-        };
-        assert_eq!(
-            channelled.total_duration(),
-            0.0,
-            "channelled has 0 duration (unbounded)"
-        );
-
-        let dot = EffectDef {
-            pulse_count: 5,
-            pulse_duration: 1.5,
-            ..Default::default()
-        };
-        assert_eq!(dot.total_duration(), 7.5, "DoT: 5 * 1.5 = 7.5");
-    }
-
-    #[test]
-    fn tcm_range_meters_known_tiers_resolve_to_expected_distances() {
-        assert_eq!(EffectDef::tcm_range_meters("Melee"), 2.5);
-        assert_eq!(EffectDef::tcm_range_meters("Short"), 5.0);
-        assert_eq!(EffectDef::tcm_range_meters("Medium"), 8.0);
-        assert_eq!(EffectDef::tcm_range_meters("Long"), 14.0);
-        assert_eq!(EffectDef::tcm_range_meters("Weapon"), 20.0);
-    }
-
-    #[test]
-    fn tcm_range_meters_unknown_tier_falls_back_to_medium() {
-        // Unknown tier defaults to Medium=8.0 with a warn log.
-        assert_eq!(EffectDef::tcm_range_meters("NoSuchTier"), 8.0);
-        assert_eq!(EffectDef::tcm_range_meters(""), 8.0);
-    }
-
-    #[test]
-    fn tcm_half_angle_radians_known_tiers_resolve() {
-        use std::f32::consts::{FRAC_PI_4, FRAC_PI_8};
-        assert!((EffectDef::tcm_half_angle_radians("Narrow") - FRAC_PI_8).abs() < 1e-6);
-        assert!((EffectDef::tcm_half_angle_radians("Medium") - FRAC_PI_4).abs() < 1e-6);
-        assert!((EffectDef::tcm_half_angle_radians("Wide") - 3.0 * FRAC_PI_8).abs() < 1e-6);
-    }
-
-    #[test]
-    fn tcm_half_angle_radians_unknown_tier_falls_back_to_medium() {
-        use std::f32::consts::FRAC_PI_4;
-        assert!((EffectDef::tcm_half_angle_radians("NoSuchTier") - FRAC_PI_4).abs() < 1e-6);
-    }
-}
